pyproject.toml
[project]
name = "deepsearchagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "attrs>=25.4.0",
    "dotenv>=0.9.9",
    "openai>=2.7.1",
    "requests>=2.32.5",
    "tavily-python>=0.7.12",
]
